# openapi.yaml
openapi: "3.0.0"
info:
  title: "AI Assistant API"
  version: "1.1.0"
  description: "API documentation for the AI Assistant application."
servers:
  - url: "https://ai-assistant-chatbot-server.vercel.app"
    description: "Production server"
  - url: "http://localhost:3000"
    description: "Local server"
...
paths:
  /api/auth/signup:
    post:
      tags:
        - "Auth"
      summary: "Sign Up a new user"
      description: "Create a new user account with the provided email and password."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
            example:
              email: "user@example.com"
              password: "yourpassword"
      responses:
        '200':
          description: "User created successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "User created successfully"
        '400':
          description: "User already exists."
        '500':
          description: "Server error."
  /api/auth/login:
    post:
      tags:
        - "Auth"
      summary: "Log in a user"
      description: "Authenticate a user with email and password and return a JWT token."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
            example:
              email: "user@example.com"
              password: "yourpassword"
      responses:
        '200':
          description: "Successfully authenticated."
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
              example:
                token: "your.jwt.token"
        '400':
          description: "Invalid credentials or user does not exist."
        '500':
          description: "Server error."
  /api/auth/verify-email:
    get:
      tags:
        - "Auth"
      summary: "Verify if an email exists"
      description: "Checks whether a user with the specified email exists."
      parameters:
        - name: email
          in: query
          description: "Email address to verify."
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: "Email verification result."
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
              example:
                exists: true
        '400':
          description: "Email is required."
        '500':
          description: "Server error."
...
  /api/auth/passkey:
    get:
      tags:
        - "Auth"
      summary: "List the authenticated user's passkeys"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "List of registered passkeys."
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: '#/components/schemas/PasskeySummary'
...
  /api/auth/reset-password:
    post:
      tags:
        - "Auth"
      summary: "Reset user password"
      description: "Resets the password for a user identified by the email."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                newPassword:
                  type: string
              required:
                - email
                - newPassword
            example:
              email: "user@example.com"
              newPassword: "newpassword123"
      responses:
        '200':
          description: "Password reset successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: "Password reset successfully"
        '400':
          description: "Email and new password are required, or user does not exist."
        '500':
          description: "Server error."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PasskeySummary:
      type: object
      properties:
        credentialID:
          type: string
          description: "Base64url-encoded credential ID. Stable across sign-ins; safe to expose."
        nickname:
          type: string
        transports:
          type: array
          items:
            type: string
        deviceType:
          type: string
          description: "Either 'singleDevice' or 'multiDevice' (synced)."
        backedUp:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
...